+++ ListableUI/Sources/Internal/PresentationState/PresentationState.HeaderFooterState.swift
@@ -48,7 +48,7 @@ protocol AnyPresentationHeaderFooterState : AnyObject
     func size(
         for info : Sizing.MeasureInfo,
         viewCache : ReusableViewCache,
-        sizingSharingCache: PresentationState.SizingSharingCache,
+        sizingSharingCache: SizingSharingCache,
         environment : ListEnvironment
     ) -> CGSize
 }
@@ -243,7 +243,7 @@ extension PresentationState
         func size(
             for info : Sizing.MeasureInfo,
             viewCache : ReusableViewCache,
-            sizingSharingCache: PresentationState.SizingSharingCache,
+            sizingSharingCache: SizingSharingCache,
             environment : ListEnvironment
         ) -> CGSize
         {
